requirements.txt
absl-py==1.3.0
aiofiles==23.2.1
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.7.0
anyio==3.7.0
apturl==0.5.2
asgiref==3.5.2
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==25.3.0
autopep8==2.0.2
awscli==1.32.3
awsebcli==3.20.10
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.3.0
beautifulsoup4==4.14.3
blessed==1.20.0
blinker==1.6.2
blis==0.2.4
botocore==1.34.3
Brlapi==0.7.0
build==1.2.2.post1
cachetools==5.2.0
cement==2.8.2
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
cloudinary==1.37.0
colorama==0.4.3
coloredlogs==15.0.1
commonmark==0.9.1
configparser==5.3.0
contourpy==1.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
dbus-python==1.2.16
decorator==5.2.1
deep-sort-realtime==1.3.2
defer==1.0.6
Deprecated==1.2.18
distlib==0.3.7
distro==1.7.0
distro-info==0.23+ubuntu1.1
Django==4.1.3
django-timescaledb==0.2.13
djangorestframework==3.14.0
dnspython==2.6.1
doqu==0.28.2
duplicity==0.8.12.0
durationpy==0.10
entrypoints==0.3
etelemetry==0.3.0
evdev==1.6.0
exceptiongroup==1.1.1
executing==1.2.0
eyantra-autoeval==0.1.17
faiss-cpu==1.8.0.post1
fastapi==0.116.1
fasteners==0.14.1
ffmpy==0.5.0
filelock==3.12.3
fitz==0.0.1.dev2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2025.3.0
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==4.8.0
gradio_client==0.7.1
graphlib_backport==1.1.0
greenlet==2.0.2
grpcio==1.70.0
gunicorn==20.1.0
h11==0.16.0
h5py==3.7.0
hf-xet==1.1.5
hidapi-usb==0.3.1
httpcore==1.0.9
httplib2==0.14.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==2.8
importlib-metadata==6.11.0
importlib-resources==5.12.0
ipython==8.6.0
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.13
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.7
kubernetes==33.1.0
langchain==0.0.354
langchain-community==0.0.20
langchain-core==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.0.87
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
looseversion==1.1.2
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.22.0
mathparse==0.1.2
matplotlib==3.7.5
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mmh3==5.0.1
monotonic==1.5
motor==3.3.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy_extensions==1.1.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.3.1
olefile==0.46
onnxruntime==1.19.2
openai==0.27.4
opencv-python==4.6.0.66
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt-einsum==3.3.0
orjson==3.10.15
overrides==7.7.0
p==1.4.0
packaging==23.2
paho-mqtt==2.1.0
pandas==1.5.1
paramiko==2.6.0
parso==0.8.3
pathlib==1.0.1
pathspec==0.10.1
pbr==5.11.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.3.0
Pint==0.21.1
pkgutil_resolve_name==1.3.10
plac==0.9.6
platformdirs==3.10.0
plotly==5.18.0
polyline==2.0.2
posthog==4.2.0
preshed==2.0.1
prettytable==3.11.0
prompt-toolkit==3.0.31
protobuf==3.20.3
prov==2.0.0
psutil==7.0.0
psycopg==3.2.3
psycopg2-binary==2.9.7
pure-eval==0.2.2
pyaadhaar==2.0.2
pyarrow==13.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.16.2
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pycups==1.9.73
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.8.1b0
pydualsense==0.7.0
pydub==0.25.1
pygame==2.4.0
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==2.6.0
pylibjpeg==1.4.0
pylibjpeg-openjpeg==1.3.2
pymacaroons==0.13.0
pymongo==4.5.0
PyMuPDF==1.22.2
PyNaCl==1.3.0
pyngrok==6.0.0
pyparsing==3.1.4
pypdf==5.8.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2022.6
pyxdg==0.26
pyxnat==1.5
PyYAML==6.0.2
pyzbar==0.1.9
rdflib==6.3.2
referencing==0.35.1
regex==2023.8.8
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
rich==12.5.1
rpds-py==0.20.1
rsa==4.7.2
ruff==0.12.4
s3transfer==0.9.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.9.3
seaborn==0.12.1
SecretStorage==2.3.1
semantic-version==2.8.5
sentence-transformers==3.0.1
sentry-sdk==1.13.0
shapely==2.0.7
shellingham==1.5.4
simplejson==3.16.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.4.3
srsly==1.0.7
stack-data==0.6.0
starlette==0.44.0
streamlit==1.28.1
sympy==1.13.3
systemd-python==234
temp==2020.7.2
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==1.1.0
testresources==2.0.1
thinc==7.0.8
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==1.13.0
torchaudio==0.8.1
torchvision==0.14.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.5.0
traits==6.3.2
transformers==4.34.1
twilio==8.1.0
typer==0.16.0
typing-inspect==0.9.0
typing_extensions==4.13.2
tzlocal==5.0.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ultralytics==8.0.13
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
uvicorn==0.26.0
uvloop==0.21.0
validators==0.22.0
virtualenv==20.24.5
wadllib==1.3.3
wasabi==0.10.1
watchdog==3.0.0
watchfiles==0.24.0
wcwidth==0.1.9
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.14.1
WTForms==3.0.1
xkit==0.0.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.20.2
